Move ViewFragmentType out of TableDisplayActivity. Remove link javadoc in Constants file.
