+++ tables_app/src/main/java/org/opendatakit/tables/data/PossibleTableViewTypes.java
@@ -21,7 +21,6 @@ import org.opendatakit.database.data.OrderedColumns;
 import org.opendatakit.database.service.DbHandle;
 import org.opendatakit.database.service.UserDbInterface;
 import org.opendatakit.exception.ServicesAvailabilityException;
-import org.opendatakit.tables.activities.TableDisplayActivity;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -43,7 +42,7 @@ public class PossibleTableViewTypes {
   private boolean mMapIsValid;
   private String mMapListFileName;
 
-  private TableDisplayActivity.ViewFragmentType mDefaultViewType;
+  private ViewFragmentType mDefaultViewType;
   private String mDetailFileName;
 
   /**
@@ -64,17 +63,17 @@ public class PossibleTableViewTypes {
     if (defaultViewType != null) {
       switch (defaultViewType) {
       case SPREADSHEET:
-        mDefaultViewType = TableDisplayActivity.ViewFragmentType.SPREADSHEET;
+        mDefaultViewType = ViewFragmentType.SPREADSHEET;
         break;
       case MAP:
-        mDefaultViewType = TableDisplayActivity.ViewFragmentType.MAP;
+        mDefaultViewType = ViewFragmentType.MAP;
         break;
       case LIST:
-        mDefaultViewType = TableDisplayActivity.ViewFragmentType.LIST;
+        mDefaultViewType = ViewFragmentType.LIST;
         break;
       }
     } else {
-      mDefaultViewType = TableDisplayActivity.ViewFragmentType.SPREADSHEET;
+      mDefaultViewType = ViewFragmentType.SPREADSHEET;
     }
 
     mSpreadsheetIsValid = true; // always
@@ -110,7 +109,7 @@ public class PossibleTableViewTypes {
   /**
    * @return the default view type for this table
    */
-  public TableDisplayActivity.ViewFragmentType getDefaultViewType() {
+  public ViewFragmentType getDefaultViewType() {
     return mDefaultViewType;
   }
