+++ tables_app/src/main/java/org/opendatakit/tables/data/ViewFragmentType.java
@@ -1,0 +1,11 @@
+package org.opendatakit.tables.data;
+
+/**
+ * The fragment types this activity could be displaying.
+ *
+ * @author sudar.sam@gmail.com
+ */
+@SuppressWarnings("JavaDoc")
+public enum ViewFragmentType {
+  SPREADSHEET, LIST, MAP, DETAIL, DETAIL_WITH_LIST, SUB_LIST
+}
