+++ tables_app/src/main/java/org/opendatakit/tables/views/webkits/OdkTables.java
@@ -20,7 +20,7 @@ import android.content.Context;
 import android.os.Bundle;
 import org.opendatakit.database.queries.BindArgs;
 import org.opendatakit.tables.activities.TableDisplayActivity;
-import org.opendatakit.tables.activities.TableDisplayActivity.ViewFragmentType;
+import org.opendatakit.tables.data.ViewFragmentType;
 import org.opendatakit.tables.utils.Constants;
 import org.opendatakit.tables.utils.IntentUtil;
 import org.opendatakit.views.ODKWebView;
